Leo bug fixes (#130)
* Refactor Chatbox and divide it into another file

* fix trashcan icon and more
